refactor(public-professionals): move persistence to infrastructure
