+++ server/features/public-professionals/README.md
@@ -1,13 +1,13 @@
 # Public Professionals — bounded context
 
-> **Tipo:** Frontera de módulo del contexto Public Professionals. **Primera
-> materialización real** del contexto en M21: la regla pura de elegibilidad del
-> banco de profesionales pasa a vivir en `domain/`.
+> **Tipo:** Frontera de módulo del contexto Public Professionals. La regla pura
+> de elegibilidad vive en `domain/` (M21) y la persistencia + mapping en
+> `infrastructure/` (M22).
 > **Origen:** [ARCH-1](../../../docs/audit/repository-domain-architecture-audit.md) · [ARCH-2](../../../docs/architecture/backend-boundary-adr.md) · [ARCH-3](../../../docs/architecture/shared-lib-boundary-inventory.md).
-> **ID:** **M21 (Fase E — apertura: dominio de elegibilidad)**.
+> **ID:** **M22 (Fase E — persistencia a `infrastructure/`)**.
 >
-> **Estado:** M21 **listo para integración** (no mergeado). **Fase E abierta.**
-> **M22 (persistencia) no iniciado.**
+> **Estado:** M21 mergeado. **M22 listo para integración** (no mergeado).
+> **Fase E abierta. M23 (ruta + rate limit) no iniciado.**
 
 Este directorio es la **frontera** del contexto Public Professionals: el
 directorio público de profesionales del portal (perfil público de clínica,
@@ -32,17 +32,26 @@ detrás de un barrel público, sin cambiar comportamiento observable.
 ## 2. Arquitectura actual
 
 ```text
-db-public-professionals → domain barrel (features/public-professionals/domain/index.ts)
+routes (public-professionals, clinic-public-profile)
+  → shim legacy (server/db-public-professionals.ts, un solo re-export)
+    → infrastructure barrel (features/public-professionals/infrastructure/index.ts)
+        ├─ public-professionals-mapping.ts     (puro; sólo tipos del shared kernel)
+        └─ public-professionals-repository.ts  → domain barrel (features/.../domain/index.ts)
 ```
 
-`server/db-public-professionals.ts` consume la elegibilidad **por el barrel
-canónico**, nunca por un archivo interno del dominio ni por el path legacy.
+- **`infrastructure/`** — persistencia y mapping movidos desde
+  `server/db-public-professionals.ts` en **M22**: `public-professionals-mapping.ts`
+  (lógica pura) + `public-professionals-repository.ts` (Drizzle, `pgClient.unsafe`,
+  SQL de elegibilidad) tras un barrel. Ver
+  [`infrastructure/README.md`](./infrastructure/README.md).
+- El **repository** consume la elegibilidad **por el domain barrel canónico**,
+  nunca por un archivo interno del dominio ni por el path legacy.
+- El **shim legacy** `server/db-public-professionals.ts` queda como un único
+  `export *` hacia el barrel de infrastructure porque las rutas todavía consumen
+  el path legacy en M22 (su reapunte/retiro es M23/M24).
 
-## 3. Qué NO mueve M21
+## 3. Qué NO mueve M22
 
-- **Persistencia** — `server/db-public-professionals.ts` sigue en su path legacy
-  (queries, templates SQL, mappings y filtros intactos byte por byte). Su
-  traslado a `infrastructure/` está reservado para **M22 (no iniciado)**.
 - **Ruta y rate limit** — `server/routes/public-professionals.fastify.ts` y
   `server/lib/public-professionals-rate-limit.ts` quedan **intactos** hasta
   **M23**.
@@ -66,21 +75,30 @@ transporte HTTP ni motor de persistencia.
 | Capa | Rol | Puede importar | No puede importar |
 | --- | --- | --- | --- |
 | **[domain](./domain/README.md)** | Reglas puras. | shared kernel (sólo tipos), utilidades puras relativas del propio contexto | `fastify`, Drizzle runtime, `env`, `http`, auth/CORS, `db-*`, `infrastructure`, `routes`, Supabase, I/O de Node, otros `server/lib/**`, `report-study-types` |
+| **[infrastructure · mapping](./infrastructure/README.md)** | Lógica pura de mapping/evaluación. | shared kernel (sólo tipos), la propia capa | DB, Drizzle runtime, `fastify`, `routes`, `application`, auth/CORS/env, rate limit, Supabase, I/O |
+| **[infrastructure · repository](./infrastructure/README.md)** | Persistencia y consultas. | `drizzle-orm`, `server/db.ts`, `drizzle/schema.ts`, el domain barrel, la propia capa | `routes`, `application`, `fastify`, auth/CORS/audit/email, Supabase, `server/lib/**`, frontend |
 
 Verificado por
-`test/architecture/public-professionals-domain-boundary-guard.test.ts`.
+`test/architecture/public-professionals-domain-boundary-guard.test.ts` (domain) y
+`test/architecture/public-professionals-infrastructure-boundary-guard.test.ts`
+(infrastructure).
 
 ## 6. Estado de la migración
 
-- **M21 — apertura de Fase E (listo para integración, no mergeado)** — move de
-  la regla de elegibilidad a `domain/` con barrel público, reapuntado del
-  consumidor runtime y del test de dominio al barrel, shim temporal en el path
-  legacy, guard de frontera nuevo y documentación. **Cero cambios** en SQL,
-  rutas, rate limits, schema, migraciones, auth ni CORS.
-- **M22 (no iniciado)** — persistencia de Public Professionals a
-  `infrastructure/`.
+- **M21 — apertura de Fase E (mergeado)** — move de la regla de elegibilidad a
+  `domain/` con barrel público, reapuntado del consumidor runtime y del test de
+  dominio al barrel, shim temporal en el path legacy `server/lib/...`, guard de
+  frontera y documentación.
+- **M22 — persistencia a `infrastructure/` (listo para integración, no
+  mergeado)** — split de `server/db-public-professionals.ts` en
+  `infrastructure/{public-professionals-mapping,public-professionals-repository}.ts`
+  tras un barrel, shim legacy de un solo re-export en el path legacy, guard de
+  infraestructura nuevo, reapunte de los tests SQL al repository canónico y del
+  guard de dominio al repository. **Cero cambios** en SQL, rutas, rate limits,
+  schema, migraciones, auth ni CORS. Ver
+  [Nota de implementación — M22](../../../docs/implementation/m22-public-professionals-infrastructure.md).
 - **M23 (no iniciado)** — ruta y rate limit.
-- **M24 (no iniciado)** — cierre de Fase E y retiro del shim legacy.
+- **M24 (no iniciado)** — cierre de Fase E y retiro de los shims legacy.
 
 ## 7. Documentos rectores
 
@@ -88,3 +106,4 @@ Verificado por
 - [ARCH-2 — Backend Domain Boundary ADR](../../../docs/architecture/backend-boundary-adr.md) — reglas de dependencia por capa.
 - [ARCH-3 — Shared / Lib Boundary Inventory](../../../docs/architecture/shared-lib-boundary-inventory.md) — inventario a nivel de archivo del terreno a migrar.
 - [Nota de implementación — M21](../../../docs/implementation/m21-public-professionals-domain.md) — mueve la regla de elegibilidad a `domain/` con shim temporal.
+- [Nota de implementación — M22](../../../docs/implementation/m22-public-professionals-infrastructure.md) — mueve la persistencia + mapping a `infrastructure/` con shim legacy.
